modules/cloud-function/hello-function/src/requirements.txt
google-cloud-secret-manager==2.10.0
SQLAlchemy==1.3.12